Move functions to actions and monkey patch variables
